fix(release): remove environment gate that blocks automated tag releases (#117)
* fix(release): remove environment gate that blocks automated tag releases

The 'environment: release' line requires manual approval which blocks
automated v* tag pushes. Remove it for fully automated releases.

Owner can add manual approval back via GitHub Settings → Environments
if desired.


* feat(mcp,core): make limits configurable via env vars

MCP server limits now configurable via environment variables:
- ATHEON_MCP_RATE_LIMIT (default: 10)
- ATHEON_MCP_RATE_BURST (default: 20)
- ATHEON_MCP_CONCURRENT_CAP (default: 50)
- ATHEON_MCP_MAX_REQUEST_BYTES (default: 64MiB)
- ATHEON_MCP_REQUEST_TIMEOUT (default: 30s)
- ATHEON_MCP_SCAN_STRING_MAX_BYTES (default: 32MiB)

Core scanner limits now configurable:
- ATHEON_SKIP_DIRS (comma-separated, appended to defaults)
- ATHEON_BINARY_EXTS (comma-separated, appended to defaults)

Also fixes Makefile lint target removing 2>/dev/null that hid errors.
Fixes test type mismatch (int64 vs int) in mcp_concurrency_test.go.


---------

+++ cmd/mcp/mcp_concurrency_test.go
@@ -105,11 +105,11 @@ func TestMCPConcurrentCapMultipleHolders(t *testing.T) {
 
 	// Give holders time to enter dispatchRequest.
 	// Spin until counter reaches cap.
-	for i := 0; i < 1000 && mcpInflight.Load() < mcpConcurrentCap; i++ {
+	for i := 0; i < 1000 && mcpInflight.Load() < int64(mcpConcurrentCap); i++ {
 		time.Sleep(time.Millisecond)
 	}
 
-	if mcpInflight.Load() < mcpConcurrentCap {
+	if mcpInflight.Load() < int64(mcpConcurrentCap) {
 		t.Skipf("holders did not enter dispatchRequest in time (counter=%d), skipping", mcpInflight.Load())
 	}
